requirements.txt
appdirs==1.4.3
asgiref==3.3.1
distlib==0.3.0
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.5.0
nose==1.3.7
psycopg2-binary==2.9.9
pygame==2.5.2
python-dateutil==2.8.1
pytz==2021.1
six==1.14.0
sqlparse==0.4.1
virtualenv==20.0.10
zipp==3.1.0
